AndresCamiloRojasVargas/dags/requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.2.2
apache-airflow-providers-common-sql==1.1.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.0
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachelib==0.9.0
cattrs==1.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
clickclick==20.10.2
cloud-init==22.2
colorama==0.4.4
colorlog==5.0.1
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
croniter==1.0.15
cryptography==3.4.8
dbus-python==1.2.18
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
filelock==3.8.0
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.2
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.2
Markdown==3.4.1
MarkupSafe==1.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.2
oauthlib==3.2.0
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==21.3
pandas==1.4.3
pendulum==2.1.2
pexpect==4.8.0
platformdirs==2.5.2
prison==0.2.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.13.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
python-dotenv==0.20.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.40
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
ssh-import-id==5.11
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
Twisted==22.1.0
ubuntu-advantage-tools==27.9
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.5
virtualenv==20.16.3
wadllib==1.3.6
Werkzeug==1.0.1
WTForms==2.3.3
zipp==1.0.0
zope.interface==5.4.0
